Extract functions to init_wsgi_middlewares  module
